refactor(TopSection): use Pascal Case `<TopSection>` `TopSection.tsx` for consistency
